+++ system/src/Grav/Framework/Flex/Interfaces/FlexCollectionInterface.php
@@ -25,9 +25,10 @@ interface FlexCollectionInterface extends ObjectCollectionInterface, NestedObjec
     /**
      * @param array $entries
      * @param FlexDirectory $directory
+     * @param string $keyField
      * @return static
      */
-    public static function createFromArray(array $entries, FlexDirectory $directory) : FlexCollectionInterface;
+    public static function createFromArray(array $entries, FlexDirectory $directory, string $keyField = null) : FlexCollectionInterface;
 
     /**
      * @param array $elements